Stream link remove prompt (#301)
* [streamlinks] Added prompt on stream link remove

* Added new line to the end of file for Toaster

* corrected format

* separate static message

* changed window name

+++ cogs/streamlinks.py
@@ -14,6 +14,7 @@ import re
 from requests.packages.urllib3.util.retry import Retry
 from requests.adapters import HTTPAdapter
 from features.reaction_context import ReactionContext
+from features.prompt import PromptSession
 
 # Pattern: "AnyText | [Subject] Page: CurrentPage / {TotalPages}"
 pagination_regex = re.compile(r'^\[([^\]]*)\]\s*Page:\s*(\d*)\s*\/\s*(\d*)')
@@ -120,8 +121,13 @@ class StreamLinks(commands.Cog):
             return
 
         stream = self.repo.get_stream_by_id(id)
-        self.repo.remove(id)
-        await ctx.reply(utils.fill_message('streamlinks_remove_success', link=stream.link))
+
+        prompt_message = utils.fill_message('streamlinks_remove_prompt', link=stream.link)
+        result = await PromptSession(self.bot, ctx, prompt_message, 60).run()
+
+        if result:
+            self.repo.remove(id)
+            await ctx.reply(utils.fill_message('streamlinks_remove_success', link=stream.link))
 
     def get_link_data(self, link: str):
         """
